Extract TraceThread class
